requirements.txt
requests
PyJWT
python-dotenv
python-telegram-bot
